Refactor: Transition to optimized chat management and enhance component structure
- Replaced useChatRoomActions with useOptimizedChatRoom for improved state management in ChatRoom.
- Updated ChatsScreen to utilize useOptimizedChats for better chat data handling and filtering.
- Refactored Avatar component to use utility functions for color and initials generation.
- Introduced new utility functions for time formatting and avatar management.
- Streamlined imports across components and services for better organization and clarity.
