+++ shared/components/Avatar.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
 import { ThemedText } from './ThemedText';
-import { User } from '@/hooks/useUser';
+import { User } from '../services/user';
+import { getAvatarColor, getInitials, getStatusColor } from '../utils/avatarUtils';
 
 interface AvatarProps {
   user?: User;
@@ -10,43 +11,10 @@ interface AvatarProps {
 }
 
 
-const getAvatarColor = (identifier?: string): string => {
-  if (!identifier) return '#C0C0C0';
-  
-  let hash = 0;
-  for (let i = 0; i < identifier.length; i++) {
-    hash = identifier.charCodeAt(i) + ((hash << 5) - hash);
-  }
-  
-  let color = '#';
-  for (let i = 0; i < 3; i++) {
-    const value = (hash >> (i * 8)) & 0xFF;
-    color += ('00' + value.toString(16)).substr(-2);
-  }
-  
-  return color;
-};
-
-const getInitials = (name?: string): string => {
-  if (!name) return '?';
-  
-  const parts = name.split(' ');
-  if (parts.length === 1) {
-    return parts[0].charAt(0).toUpperCase();
-  }
-  
-  return (parts[0].charAt(0) + parts[parts.length - 1].charAt(0)).toUpperCase();
-};
 
 export function Avatar({ user, size = 40, showStatus = true }: AvatarProps) {
   const backgroundColor = getAvatarColor(user?.id || user?.name);
   const initials = getInitials(user?.name);
-  
-  const statusColors = {
-    online: '#4CAF50',
-    offline: '#9E9E9E',
-    away: '#FFC107',
-  };
 
   return (
     <View style={styles.container}>
@@ -68,7 +36,7 @@ export function Avatar({ user, size = 40, showStatus = true }: AvatarProps) {
           style={[
             styles.statusIndicator,
             {
-              backgroundColor: statusColors[user.status],
+              backgroundColor: getStatusColor(user.status),
               width: size / 4,
               height: size / 4,
               borderRadius: size / 8,
