+++ shared/utils/timeUtils.ts
@@ -1,0 +1,39 @@
+/**
+ * Utility functions for time formatting and manipulation
+ */
+
+export const formatTime = (timestamp: number): string => {
+  const date = new Date(timestamp);
+  return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
+};
+
+export const formatChatTime = (timestamp: number): string => {
+  const date = new Date(timestamp);
+  const now = new Date();
+  const diffInDays = Math.floor(
+    (now.getTime() - date.getTime()) / (1000 * 60 * 60 * 24)
+  );
+
+  if (diffInDays === 0) {
+    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
+  } else if (diffInDays === 1) {
+    return "Yesterday";
+  } else if (diffInDays < 7) {
+    return date.toLocaleDateString([], { weekday: "short" });
+  } else {
+    return date.toLocaleDateString([], { month: "short", day: "numeric" });
+  }
+};
+
+export const isToday = (timestamp: number): boolean => {
+  const date = new Date(timestamp);
+  const today = new Date();
+  return date.toDateString() === today.toDateString();
+};
+
+export const isYesterday = (timestamp: number): boolean => {
+  const date = new Date(timestamp);
+  const yesterday = new Date();
+  yesterday.setDate(yesterday.getDate() - 1);
+  return date.toDateString() === yesterday.toDateString();
+};
